feat: Melee attack state, public mouse direction
The melee attack state script was created, initialized in the player, and recommendations for its use were commented in the script.
A public read-only variable was created with the mouse direction for the animator to read.
Comments were left in the Idel and Walk states with instructions for implementing the melee attack.